Squashed Azure storage commits
