+++ requirements-server.txt
@@ -21,6 +21,10 @@ autobahn==21.11.1
     # via daphne
 automat==20.2.0
     # via twisted
+azure-core==1.21.1
+    # via azure-storage-blob
+azure-storage-blob==12.9.0
+    # via django-storages
 billiard==3.6.4.0
     # via celery
 boto3==1.20.20
@@ -32,7 +36,9 @@ botocore==1.23.20
 celery==5.2.2
     # via -r requirements-server.in
 certifi==2021.10.8
-    # via requests
+    # via
+    #   msrest
+    #   requests
 cffi==1.15.0
     # via cryptography
 chainmap==1.0.3
@@ -68,6 +74,7 @@ coreschema==0.0.4
 cryptography==36.0.0
     # via
     #   autobahn
+    #   azure-storage-blob
     #   pyopenssl
     #   service-identity
 daphne==3.0.2
@@ -78,6 +85,7 @@ django==3.2.12
     # via
     #   -r requirements-server.in
     #   channels
+    #   django-debug-toolbar
     #   django-filter
     #   django-model-utils
     #   django-request-logging
@@ -87,13 +95,15 @@ django==3.2.12
     #   drf-yasg
 django-cleanup==5.2.0
     # via -r requirements-server.in
+django-debug-toolbar==3.2.4
+    # via -r requirements-server.in
 django-filter==21.1
     # via -r requirements-server.in
 django-model-utils==4.2.0
     # via -r requirements-server.in
 django-request-logging==0.7.3
     # via -r requirements-server.in
-django-storages==1.12.3
+django-storages[azure]==1.12.3
     # via -r requirements-server.in
 djangorestframework==3.12.4
     # via
@@ -123,6 +133,8 @@ incremental==21.3.0
     # via twisted
 inflection==0.5.1
     # via drf-yasg
+isodate==0.6.1
+    # via msrest
 itypes==1.2.0
     # via coreapi
 jinja2==3.0.3
@@ -141,10 +153,14 @@ markdown==3.3.6
     # via -r requirements-server.in
 markupsafe==2.0.1
     # via jinja2
+msrest==0.6.21
+    # via azure-storage-blob
 numpy==1.21.4
     # via
     #   pandas
     #   pyarrow
+oauthlib==3.2.0
+    # via requests-oauthlib
 ods-tools==2.1.0.post1
     # via -r requirements-server.in
 packaging==21.3
@@ -187,11 +203,15 @@ pytz==2021.3
     #   django
     #   pandas
 requests==2.26.0
-    # via coreapi
-ruamel-yaml==0.17.17
+    # via
+    #   azure-core
+    #   coreapi
+    #   msrest
+    #   requests-oauthlib
+requests-oauthlib==1.3.1
+    # via msrest
+ruamel.yaml==0.17.17
     # via drf-yasg
-ruamel-yaml-clib==0.2.6
-    # via ruamel.yaml
 s3transfer==0.5.0
     # via boto3
 service-identity==21.1.0
@@ -199,7 +219,9 @@ service-identity==21.1.0
 six==1.16.0
     # via
     #   automat
+    #   azure-core
     #   click-repl
+    #   isodate
     #   pathlib2
     #   pyopenssl
     #   python-dateutil
@@ -207,7 +229,9 @@ six==1.16.0
 sqlalchemy==1.4.27
     # via -r requirements-server.in
 sqlparse==0.4.2
-    # via django
+    # via
+    #   django
+    #   django-debug-toolbar
 twisted[tls]==21.7.0
     # via daphne
 txaio==21.2.1
