+++ requirements.txt
@@ -30,7 +30,13 @@ autobahn==21.11.1
     # via daphne
 automat==20.2.0
     # via twisted
-backports-entry-points-selectable==1.1.1
+azure-core==1.21.1
+    # via azure-storage-blob
+azure-storage-blob==12.9.0
+    # via django-storages
+backcall==0.2.0
+    # via ipython
+backports.entry-points-selectable==1.1.1
     # via virtualenv
 backports-tempfile==1.0
     # via -r requirements.in
@@ -57,6 +63,7 @@ celery==5.2.1
 certifi==2021.10.8
     # via
     #   fiona
+    #   msrest
     #   pyproj
     #   requests
 cffi==1.15.0
@@ -118,6 +125,7 @@ cramjam==2.5.0
 cryptography==36.0.0
     # via
     #   autobahn
+    #   azure-storage-blob
     #   pyopenssl
     #   service-identity
 daphne==3.0.2
@@ -149,7 +157,7 @@ django-model-utils==4.2.0
     # via -r ./requirements-server.in
 django-request-logging==0.7.4
     # via -r ./requirements-server.in
-django-storages==1.12.3
+django-storages[azure]==1.12.3
     # via -r ./requirements-server.in
 django-webtest==1.9.8
     # via -r requirements.in
@@ -240,6 +248,8 @@ model-mommy==2.0.0
     # via -r requirements.in
 msgpack==1.0.3
     # via oasislmf
+msrest==0.6.21
+    # via azure-storage-blob
 munch==2.5.0
     # via fiona
 numba==0.54.1
@@ -259,6 +269,8 @@ numpy==1.20.3
     #   scipy
 oasislmf[extra]==1.25.1
     # via -r ./requirements-worker.in
+oauthlib==3.2.0
+    # via requests-oauthlib
 ods-tools==2.0.1
     # via -r ./requirements-server.in
 packaging==21.3
@@ -356,10 +368,15 @@ pytz==2021.3
 requests==2.26.0
     # via
     #   -r requirements.in
+    #   azure-core
     #   cookiecutter
     #   coreapi
+    #   msrest
     #   oasislmf
+    #   requests-oauthlib
     #   requests-toolbelt
+requests-oauthlib==1.3.1
+    # via msrest
 requests-toolbelt==0.9.1
     # via oasislmf
 rtree==0.9.7
@@ -386,10 +403,12 @@ six==1.16.0
     # via
     #   anytree
     #   automat
+    #   azure-core
     #   click-repl
     #   cookiecutter
     #   fasteners
     #   fiona
+    #   isodate
     #   munch
     #   pathlib2
     #   pyopenssl
