+++ src/server/oasisapi/settings.py
@@ -208,10 +208,41 @@ AWS_S3_OBJECT_PARAMETERS = {
 }
 
 
+# Required Blob storage settings
+# https://stackoverflow.com/questions/54729137/django-azure-upload-file-to-blob-storage/54767932
+"""
+AZURE_ACCOUNT_NAME = '<azure storage account name>'
+AZURE_ACCOUNT_KEY = '<azure storage account key for this container>'
+AZURE_CONTAINER = '<blob container name>'
+AZURE_LOCATION = '<subdir in blob container name>'
+"""
+AZURE_ACCOUNT_NAME = iniconf.settings.get('server', 'AZURE_ACCOUNT_NAME', fallback=None)
+AZURE_ACCOUNT_KEY  = iniconf.settings.get('server', 'AZURE_ACCOUNT_KEY', fallback=None)
+AZURE_CUSTOM_DOMAIN = f'{AZURE_ACCOUNT_NAME}.blob.core.windows.net'
+AZURE_CONTAINER = iniconf.settings.get('server', 'AZURE_CONTAINER', fallback=None)
+AZURE_LOCATION = iniconf.settings.get('server', 'AZURE_LOCATION', fallback='')
+AZURE_SHARED_CONTAINER = iniconf.settings.get('server', 'AZURE_SHARED_CONTAINER', fallback=True)
+AZURE_OVERWRITE_FILES = iniconf.settings.get('server', 'AZURE_OVERWRITE_FILES', fallback=True)
+
+## Optional Blob storage settings
+# WARNING, adding default settings with 'None' casues storage adapter to break 
+#AZURE_SSL = iniconf.settings.get('server', 'AZURE_SSL', fallback=True)
+#AZURE_UPLOAD_MAX_CONN = iniconf.settings.get('server', 'AZURE_UPLOAD_MAX_CONN', fallback=2)
+#AZURE_CONNECTION_TIMEOUT_SECS = iniconf.settings.get('server', 'AZURE_CONNECTION_TIMEOUT_SECS', fallback=20)
+#AZURE_BLOB_MAX_MEMORY_SIZE = iniconf.settings.get('server', 'AZURE_BLOB_MAX_MEMORY_SIZE', fallback='2MB')
+#AZURE_URL_EXPIRATION_SECS = iniconf.settings.get('server', 'AZURE_URL_EXPIRATION_SECS', fallback=None)
+#AZURE_CONNECTION_STRING = iniconf.settings.get('server', 'AZURE_CONNECTION_STRING', fallback=None)
+#AZURE_TOKEN_CREDENTIAL = iniconf.settings.get('server', 'AZURE_TOKEN_CREDENTIAL', fallback=None)
+#AZURE_CACHE_CONTROL = iniconf.settings.get('server', 'AZURE_CACHE_CONTROL', fallback=None)
+#AZURE_OBJECT_PARAMETERS = iniconf.settings.get('server', 'AZURE_OBJECT_PARAMETERS', fallback=None)
+
+
+
 # Select Data Storage
 STORAGE_TYPE = iniconf.settings.get('server', 'storage_type', fallback="").lower()
 LOCAL_FS = ['local-fs', 'shared-fs']
 AWS_S3 = ['aws-s3', 's3', 'aws']
+AZURE = ['azure']
 
 if STORAGE_TYPE in LOCAL_FS:
     # Set Storage to shared volumn mount
@@ -221,6 +252,9 @@ elif STORAGE_TYPE in AWS_S3:
     # AWS S3 Object Store via `Django-Storages`
     DEFAULT_FILE_STORAGE = 'storages.backends.s3boto3.S3Boto3Storage'
     set_aws_log_level(AWS_LOG_EVEL)
+
+elif STORAGE_TYPE in AZURE:
+    DEFAULT_FILE_STORAGE = 'storages.backends.azure_storage.AzureStorage'
 
 else:
     raise ImproperlyConfigured('Invalid value for STORAGE_TYPE: {}'.format(STORAGE_TYPE))
